Move multiplication to a new mul.rs
